Refactored the state management quite a bit
 * Did away with the concept of an `IServiceProvider` being suplied to
the `TurnContextServiceCollect` for upstream service resolution and
living with fact that we have no DI for now
 * StateManagementMiddleware responsibilities extended to manage
`IStateStores`, `IStateManagers` and the mappings between them.
 * Introduced single `IStateManagerResolver` service into the
`ITurnContext::Services` now and both the middleware loading/saving
logic as well as the `ITurnContext` extension methods now work with
that to get state managers instead of looking for them directly in the
`ITurnContext`'s `Services` collection.
 * Updated Alarm samples to use new StateManagementMiddleware
configuration APIs